--- haystack/preview/llm_backends/chat_message.py
@@ -1,7 +1,0 @@
-from dataclasses import dataclass
-
-
-@dataclass
-class ChatMessage:
-    content: str
-    role: str
